backend/requirements.txt
# LevantiLearn Backend — Core Dependencies
# Install with: pip install -r requirements.txt
# Run with:     python -m uvicorn main:app --reload

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Auth — Supabase (replaces Firebase)
supabase==2.4.0
python-jose[cryptography]==3.3.0

# Caching
redis==5.0.4
hiredis==2.3.2

# Data utilities
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4

# TTS + STT — OpenAI API (tts-1-hd for Arabic, whisper-1 for transcription)
openai>=1.30.0
# edge-tts kept as fallback if no OPENAI_API_KEY is set
edge-tts>=6.1.9

# HTTP client — pinned to satisfy supabase==2.4.0 (<0.26) and fastapi (>=0.23)
httpx>=0.24,<0.26

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9

# Dev / testing
pytest==8.2.0
pytest-asyncio==0.23.7
